requirements.txt
beautifulsoup4==4.12.2
dominate==2.7.0
matplotlib==3.7.1
natsort==8.3.1
numpy==1.24.3
onnx==1.14.0
onnxruntime==1.14.1
Pillow==9.5.0
Requests~=2.28.1
Shapely==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
visdom==0.2.4

scikit-image~=0.20.0
